+++ src/main/java/Command/RestrictedModule/ShutDownCommand.java
@@ -12,6 +12,7 @@ import Setting.Prefix;
 import Main.*;
 import Utility.SmartLogger;
 import java.awt.Color;
+import java.io.IOException;
 import java.time.Instant;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -58,12 +59,14 @@ public class ShutDownCommand implements Command{
 
                 try {
                     Thread.sleep(2000);
+                    
+                    SmartLogger.updateLog("Bot Shut Down Attemp");
+                    Main.shutdown();
                 } catch (InterruptedException ite) {
                     SmartLogger.errorLog(ite, e, this.getClass().getName(), "Thread Sleep process interrupted.");
+                } catch (IOException ex) {
+                    Logger.getLogger(ShutDownCommand.class.getName()).log(Level.SEVERE, null, ex);
                 }
-                
-                SmartLogger.updateLog("Bot Shut Down Attemp");
-                Main.shutdown();
             }
             else
                 e.getChannel().sendMessage(Emoji.error + " This command is for **Bot Owner** only!").queue();