+++ src/main/java/Resource/Emoji.java
@@ -105,8 +105,12 @@ public class Emoji {
     //Music
     public static String globe = EmojiParser.parseToUnicode(":globe_with_meridians:");
     public static String notes = EmojiParser.parseToUnicode(":notes:");
+    public static String pause = EmojiParser.parseToUnicode(":black_right_pointing_triangle_with_double_vertical_bar:");
+    public static String resume = EmojiParser.parseToUnicode(":arrow_forward:");
+    public static String stop = EmojiParser.parseToUnicode(":black_square_for_stop:");
     public static String volume_low = EmojiParser.parseToUnicode(":sound:");
     public static String volume_high = EmojiParser.parseToUnicode(":loud_sound:");
+    
     
     //Fun
     //-Game